+++ src/um/shobjidl.rs
@@ -11,14 +11,11 @@ use shared::minwindef::{BOOL, DWORD, UINT};
 use shared::windef::HWND;
 use um::objidl::IBindCtx;
 use um::propkeydef::REFPROPERTYKEY;
-use um::propsys::{GETPROPERTYSTOREFLAGS, IPropertyDescriptionList, IPropertyStore};
+use um::propsys::{IPropertyDescriptionList, IPropertyStore, GETPROPERTYSTOREFLAGS};
 use um::shobjidl_core::{IModalWindow, IModalWindowVtbl, IShellItem, IShellItemFilter, SFGAOF};
 use um::shtypes::COMDLG_FILTERSPEC;
 use um::unknwnbase::{IUnknown, IUnknownVtbl};
 use um::winnt::{HRESULT, LPCWSTR, LPWSTR, WCHAR};
-
-DEFINE_GUID!{CLSID_TaskbarList,
-    0x56fdf344, 0xfd6d, 0x11d0, 0x95, 0x8a, 0x00, 0x60, 0x97, 0xc9, 0xa0, 0x90}
 
 pub type IFileOperationProgressSink = IUnknown; // TODO
 ENUM!{enum SIATTRIBFLAGS {
@@ -384,32 +381,4 @@ interface IFileDialog2(IFileDialog2Vtbl): IFileDialog(IFileDialogVtbl) {
     fn SetNavigationRoot(
         psi: IShellItem,
     ) -> HRESULT,
-});
-
-RIDL!(
-#[uuid(0x56fdf342, 0xfd6d, 0x11d0, 0x95, 0x8a, 0x00, 0x60, 0x97, 0xc9, 0xa0, 0x90)]
-interface ITaskbarList(ITaskbarListVtbl): IUnknown(IUnknownVtbl) {
-    fn HrInit() -> HRESULT,
-    fn AddTab(
-        hwnd: HWND,
-    ) -> HRESULT,
-    fn DeleteTab(
-        hwnd: HWND,
-    ) -> HRESULT,
-    fn ActivateTab(
-        hwnd: HWND,
-    ) -> HRESULT,
-    fn SetActiveAlt(
-        hwnd: HWND,
-    ) -> HRESULT,
-});
-
-
-RIDL!(
-#[uuid(0x602d4995, 0xb13a, 0x429b, 0xa6, 0x6e, 0x19, 0x35, 0xe4, 0x4f, 0x43, 0x17)]
-interface ITaskbarList2(ITaskbarList2Vtbl): ITaskbarList(ITaskbarListVtbl) {
-    fn MarkFullscreenWindow(
-        hwnd: HWND,
-        fFullscreen: BOOL,
-    ) -> HRESULT,    
 });
